+++ packages/web3-rpc-providers/CHANGELOG.md
@@ -43,12 +43,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0.rc.1]
 
- ### Added
+### Added
 
  - When error is returned with code 429, throw rate limit error (#7102)
 
- ### Changed
+### Changed
 
  - Change request return type `Promise<ResultType>` to `Promise<JsonRpcResponseWithResult<ResultType>>` (#7102)
 
 ## [Unreleased]
+
+### Added
+
+-   Updated rate limit error of QuickNode provider for HTTP transport